driver: src: global: Added aie2ps register definitions
Added a new file xaie2psgbl_params.h with the register definitions for AIE2PS.
This file is a combination of the following header files from AIE2PS
register database r0p7:
    + me_core_module.h
    + aie2ps_mem_tile_module.h
    + aie2ps_memory_module.h
    + aie2ps_noc_module.h
    + aie2ps_pl_module.h
    + aie2ps_pl_north_module.h

In addition to adding the headers, the following changes are made to the
register definitions for some cleanup.
    + AIE2PS is replaced with XAIE2PSGBL
    + Replaced value__<MACRO> defs with <MACRO>
    + Replaced <__> with "_"
    + Deleted lines similar to-
        // Register AIE2PS_MEMORY_MODULE_PERFORMANCE_CONTROL0
    + Removed unnecessary comments at the end of some lines(/* DATA */)

These changes are made to keep the register definitions consistent with
first generation AIE register definitions.

Signed-off-by: Sankarji Gopalakrishnan <Sankarji@amd.com>

Acked-for-series: Tejus Siddagangaiah <tejus.siddagangaiah@amd.com>