src: Fix Init and Reset CDO mismatch (#46)
In Timer Sync function, no need to clear Broadcast
configuration for second broadcast channel as it
was not configured. So removed function to clear
Broadcast configuration for second channel.
Also For AIE1 Reset CDO was differ because of Isolation
was not happening for all tiles. Which has been fix in
this PR.

Signed-off-by: Keyur Parekh <Keyur.Parekh@amd.com>